requirements.txt
google-cloud-vision==3.4.4
supabase==2.3.4
python-dotenv==1.0.0
Pillow==10.1.0
